Move contents of __init__.py into base.py
Finish refactor of get_primary_keys and get_pk_constraint to improve
readability.
