+++ src/databricks/sqlalchemy/_parse.py
@@ -1,10 +1,38 @@
-from typing import List, Optional
+from typing import List, Optional, Dict
 import re
+
+from sqlalchemy.engine import CursorResult
 
 """
 This module contains helper functions that can parse the contents
-of DESCRIBE TABLE EXTENDED calls. Mostly wrappers around regexes.
+of metadata and exceptions received from DBR. These are mostly just
+wrappers around regexes.
 """
+
+def _match_table_not_found_string(message: str) -> bool:
+    """Return True if the message contains a substring indicating that a table was not found"""
+
+    DBR_LTE_12_NOT_FOUND_STRING = "Table or view not found"
+    DBR_GT_12_NOT_FOUND_STRING = "TABLE_OR_VIEW_NOT_FOUND"
+    return any(
+        [
+            DBR_LTE_12_NOT_FOUND_STRING in message,
+            DBR_GT_12_NOT_FOUND_STRING in message,
+        ]
+    )
+
+
+def _describe_table_extended_result_to_dict_list(result: CursorResult) -> List[Dict[str, str]]:
+    """Transform the CursorResult of DESCRIBE TABLE EXTENDED into a list of Dictionaries
+    """
+
+    rows_to_return = []
+    for row in result:
+        this_row = {"col_name": row.col_name, "data_type": row.data_type}
+        rows_to_return.append(this_row)
+
+    return rows_to_return
+
 
 def extract_identifiers_from_string(input_str: str) -> List[str]:
     """For a string input resembling (`a`, `b`, `c`) return a list of identifiers ['a', 'b', 'c']"""
@@ -143,3 +171,76 @@ def build_fk_dict(
     }
 
     return complete_foreign_key_dict
+
+def _parse_pk_columns_from_constraint_string(constraint_str: str) -> List[str]:
+    """Build a list of constrained columns from a constraint string returned by DESCRIBE TABLE EXTENDED
+
+    For example:
+
+    PRIMARY KEY (`id`, `name`, `email_address`)
+
+    Returns a list like
+
+    ["id", "name", "email_address"]
+    """
+
+    _extracted = extract_identifiers_from_string(constraint_str)
+
+    return _extracted
+
+def build_pk_dict(pk_name: str, pk_constraint_string: str) -> dict:
+    """Given a primary key name and a primary key constraint string, return a dictionary
+    with the following keys:
+    
+    constrained_columns
+      A list of string column names that make up the primary key
+
+    name
+      The name of the primary key constraint
+    """
+    
+    constrained_columns = _parse_pk_columns_from_constraint_string(pk_constraint_string)
+
+    return {"constrained_columns": constrained_columns, "name": pk_name}
+    
+def match_dte_rows_by_value(dte_output: List[Dict[str, str]], match: str) -> List[dict]:
+    """Return a list of dictionaries containing only the col_name:data_type pairs where the `data_type`
+    value contains the match argument.
+
+    Today, DESCRIBE TABLE EXTENDED doesn't give a deterministic name to the fields
+    a constraint will be found in its output. So we cycle through its output looking
+    for a match. This is brittle. We could optionally make two roundtrips: the first
+    would query information_schema for the name of the constraint on this table, and
+    a second to DESCRIBE TABLE EXTENDED, at which point we would know the name of the
+    constraint. But for now we instead assume that Python list comprehension is faster
+    than a network roundtrip
+    """
+
+    output_rows = []
+
+    for row_dict in dte_output:
+        if match in row_dict["data_type"]:
+            output_rows.append(row_dict)
+        
+    return output_rows
+
+def get_fk_strings_from_dte_output(dte_output: List[List]) -> List[dict]:
+    """If the DESCRIBE TABLE EXTENDED output contains foreign key constraints, return a list of dictionaries,
+    one dictionary per defined constraint
+    """
+
+    output = match_dte_rows_by_value(dte_output, "FOREIGN KEY")
+
+    return output
+
+    
+def get_pk_strings_from_dte_output(dte_output: List[Dict[str, str]]) -> Optional[List[dict]]:
+    """If the DESCRIBE TABLE EXTENDED output contains primary key constraints, return a list of dictionaries,
+    one dictionary per defined constraint.
+
+    Returns None if no primary key constraints are found.
+    """
+
+    output = match_dte_rows_by_value(dte_output, "PRIMARY KEY")
+
+    return output
